requirements.txt
numpy==2.1.3
asgiref==3.8.1 
django==5.1.3 
djangorestframework